feat: refactor country data fetching and rendering, add playground page for layout testing
Refactor Layout for Semantic HTML, Accessibility & Responsive Usability
Fixes #8
