+++ src/pages/index.html
@@ -5,6 +5,7 @@
         <i class="bi bi-globe-americas"></i>
         World Countries Directory
     </h1>
+    <hr class="border border-success border-2 opacity-50">
     <form class="row my-3" id="search-form">
         <div class="col">
             <input class="form-control" type="text" name="in-country" id="in-country">
@@ -16,23 +17,21 @@
             </button>
         </div>
     </form>
-    <div class="row my-3">
-        <table class="table table-striped table-hover">
-            <thead>
-                <tr>
-                    <th>Flag</th>
-                    <th>Name</th>
-                    <th class="text-end">Area (km<sup>2</sup>)</th>
-                    <th class="text-end">Population</th>
-                    <th>Capital</th>
-                    <th>Continents</th>
-                    <th>Region</th>
-                    <th>Subregion</th>
-                </tr>
-            </thead>
-            <tbody class="table-group-divider" id="country-list"></tbody>
-        </table>
-    </div>
+    <table class="table table-striped table-hover my-3">
+        <thead>
+            <tr>
+                <th>Flag</th>
+                <th>Name</th>
+                <th class="text-end">Area (km<sup>2</sup>)</th>
+                <th class="text-end">Population</th>
+                <th>Capital</th>
+                <th>Continents</th>
+                <th>Region</th>
+                <th>Subregion</th>
+            </tr>
+        </thead>
+        <tbody class="table-group-divider" id="country-list"></tbody>
+    </table>
 {% endblock %}
 
 {% block scripts %}
